requirement.txt
openpyxl==3.1.5
selenium==4.28.1
pandas==2.2.3
python-dotenv==1.0.1